backend/requirements.txt
django==5.0.7
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
psycopg[binary]
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
whitenoise==6.7.0
gunicorn==22.0.0
python-dateutil==2.9.0.post0
